pdf_utils.py
import fitz  # PyMuPDF

def load_pdf_chunks(pdf_path, chunk_size=300):
    doc = fitz.open(pdf_path)
    chunks = []
    for page in doc:
        text = page.get_text()
        for i in range(0, len(text), chunk_size):
            chunk = text[i:i+chunk_size].strip()
            if chunk:
                chunks.append(chunk)
    return chunks