requirements.txt
streamlit
faiss-cpu==1.7.4
sentence-transformers
gpt4all
PyMuPDF
